fix dat.GUI button elements layout
without this fix, buttons added to the GUI with a long name were artificially truncated to the same length that a slider name gets truncated (to make room for the slider and text input).  The button should not have that restriction.

After this fix, the complete long name of the button is displayed properly.