Accept embedded gltf files as _meshfile_objects
This accepts embedded .gltf files as _meshfile_objects. It may also
accept distributed gltf files as long as the meshcat server respects
meshcat's requests for the missing .bin and image files.

This cannot be part of the ExtensibleObjectLoader (EOL). EOL inhertis
from ObjectLoader where as GLTFLoader inherits from Loader. The workflow
of ObjectLoader is incompatible with GLTFLoader's workflow. Fortunately,
GLTFLoader produces a THREE.Group that we can simply stash in the scene.

To support PBR materials (reflective surfaces which reflect an environment
map), we change the "background" gradient to support environment textures.
This is largely an innocuous change. However, there is a distinct
difference between the new background and the old.

In the old, the gradient was registered in *screen* space. Whether I was
looking up or down, the top color was top of the *screen* and the bottom
at the bottom of the *screen*.

The new environment map now use the "top" (+z infinitely far away) and
bottom (-z infinitely far away) of the *world*. So, rotating the camera
to look up sees the upper pole and looking down shows the lower pole.