glTF files now cast and receive shadows
Previously, the cast/receive shadow properties were only set on the *root*
of the glTF scene -- a three.js Group node. Because the glTF object can
comprise an arbitrary hierarchy of geometry, each of the meshes in the
hierarchy need to be individually configured, otherwise they won't
participate in shadows.

Update the gltf test:
 - gltf now has the ability to cast shadows on itself.
 - Shadows are automatically enabled.