Fix setting background visibility
Introducing the environment map had introduced a bug in background
visibility. Toggling the "Background" *folder (the GUI element)
successfully toggled background visibility. But calling set_property()
on it, didn't work.

This changes how the folder is defined so that it is still valid to call
`set_property("\Background", "visible", bool)`.