Implemented VR and XR support with a controller visualizer for WebXR controllers.
added VR and XR support

Correct initial headset view

In starting a session, we create a reference space such that the headset
is at the same location in the scene as the non-VR camera's position.
Furthermore, the space turns so that the headset is looking in the same
heading as the non-VR camera was. It's not *fully* aligned, we don't want
to turn looking forward into looking down. We make sure that the scene's
up frame is still parallel with the direction of Earth's gravity.

Incidentally beef up the test html to provide testing and information about
more of the functionality.