Attempt to allow interactive camera with stability
The change events mean that I may get madness. The end event usually comes
with a good pose. However, change events may be processed after the end,
so, this attempts to marry both worlds. Get some degree of interactivity
and get a good pose at the end of the interaction.

It appears that fast mouse movement exacerbates the madness.