Allow registration of callback when camera pose gets changed
The callback is given as a string (which gets evaluated into a function).
The callback receives the viewer as its only parameter, so it has access
the viewer's APIs. The callback can be used to echo the camera position.

The callback is only invoked upon the completion of an interaction -- i.e.
upon a mouse *up* event. Calling it on every change can cause event
handling race conditions where the camera matrix is not in a stable state.

To see it at work, see the camera_callback.html.

# Conflicts:
#	dist/main.min.js