Add webpack rebuild via Dockerfile
This provides a hermetic rebuild environment for main.min.js rebuilds
destined for master, to have a single canonical answer for what the
minified content should look like. (We should probably add a GitHub
action to enforce this on pull requests.)

This also enables Drake Developers to develop changes locally without
polluting their computer due to the existing setup instructions' lack
of specifics for how and where to install node.

Also, fix gitignore snafus with an unwanted license file -- shrapnel
due to the lack of tooling.